Prometheus: Add matrix benchmark
